Issue 5733/1: int->size_t in make_script_from_event ()
... and similarly in Fingering_engraver::make_script ().
This fixes some warnings.

+++ lily/include/script-interface.hh
@@ -41,7 +41,7 @@ public:
 };
 
 void make_script_from_event (Grob *p, Context *tg,
-                             SCM type, int index);
+                             SCM type, size_t index);
 
 #endif /* SCRIPT_INTERFACE_HH */
